Refactor portal types into new portal-types.ts module
