+++ src/lib/auth.ts
@@ -4,9 +4,10 @@ import { AppMode } from "../models/stores/stores";
 import { QueryParams, DefaultUrlParams } from "../utilities/url-params";
 import { NUM_FAKE_STUDENTS, NUM_FAKE_TEACHERS } from "../components/demo/demo-creator";
 import { AppConfigModelType } from "../models/stores/app-config-model";
-import { IPortalClassOffering } from "../models/stores/user";
+import { IPortalClassOffering, UserType } from "../models/stores/user";
 import { getErrorMessage } from "../utilities/super-agent-helpers";
 import { getPortalOfferings, getPortalClassOfferings,  getProblemIdForAuthenticatedUser } from "./portal-api";
+import { PortalJWT, PortalFirebaseJWT } from "./portal-types";
 import { Logger, LogEventName } from "../lib/logger";
 import initials from "initials";
 
@@ -102,24 +103,6 @@ export interface AuthQueryParams {
   domain?: string;
 }
 
-export type PortalJWT = PortalStudentJWT | PortalTeacherJWT | PortalUserJWT;
-
-export interface BasePortalJWT {
-  alg: string;
-  iat: number;
-  exp: number;
-  uid: number;
-}
-
-export interface PortalStudentJWT extends BasePortalJWT {
-  domain: string;
-  user_type: "learner";
-  user_id: string;
-  learner_id: number;
-  class_info_url: string;
-  offering_id: number;
-}
-
 // An explicitly set appMode takes priority
 // Otherwise, assume that local users are devs, unless a token is specified,
 // in which authentication is likely being tested
@@ -128,71 +111,6 @@ export const getAppMode = (appModeParam?: AppMode, token?: string, host?: string
            ? appModeParam
            : (token == null && (host === "localhost" || host === "127.0.0.1") ? "dev" : "authed");
 };
-
-export type UserType = "student" | "teacher";
-
-export interface PortalTeacherJWT extends BasePortalJWT {
-  domain: string;
-  user_type: "teacher";
-  user_id: string;
-  teacher_id: number;
-}
-
-export interface PortalUserJWT extends BasePortalJWT {
-  domain: string;
-  user_type: "user";
-  user_id: string;
-  first_name: string;
-  last_name: string;
-}
-
-// firebase JWT claims are available to firestore security rules under request.auth.token
-export interface PortalFirebaseJWTBaseClaims {
-  user_id: string;          // e.g. `https://learn.concord.org/users/${platform_user_id}`
-  class_hash: string;
-  platform_id: string;      // e.g. "https://learn.concord.org"
-  platform_user_id: number;
-}
-
-export interface PortalFirebaseJWTStudentClaims extends PortalFirebaseJWTBaseClaims {
-  user_type: "learner";
-  offering_id: number;
-}
-
-export interface PortalFirebaseJWTTeacherClaims {
-  user_type: "teacher";
-}
-
-export type PortalFirebaseJWTClaims = PortalFirebaseJWTStudentClaims | PortalFirebaseJWTTeacherClaims;
-
-export interface BasePortalFirebaseJWT {
-  alg: string;
-  iss: string;
-  sub: string;
-  aud: string;
-  iat: number;
-  exp: number;
-  uid: string;
-  user_id: string;
-}
-
-export interface PortalFirebaseStudentJWT extends BasePortalFirebaseJWT {
-  domain: string;
-  domain_uid: number;
-  externalId: number;
-  returnUrl: string;
-  logging: boolean;
-  class_info_url: string;
-  claims: PortalFirebaseJWTStudentClaims;
-}
-
-export interface PortalFirebaseTeacherJWT extends BasePortalFirebaseJWT {
-  domain: string;
-  domain_uid: number;
-  claims: PortalFirebaseJWTTeacherClaims;
-}
-
-export type PortalFirebaseJWT = PortalFirebaseStudentJWT | PortalFirebaseTeacherJWT;
 
 export const getPortalJWTWithBearerToken = (basePortalUrl: string, type: string, rawToken: string) => {
   return new Promise<[string, PortalJWT]>((resolve, reject) => {
